# shared/openapi.yaml
openapi: 3.1.1
info:
  title: Toolgate Remote Core API
  version: 0.1.0
  description: Normative draft contract shared by Go, Rust and the TypeScript SDK. Business execution
    remains in the provider backend. This file specifies a product to implement; it is not a deployed
    API.
servers:
- url: https://toolgate.example.invalid
  description: Documentation placeholder; replace only through deployment config.
security:
- ServiceKey: []
...
paths:
  /v1/workflows:
    post:
      operationId: prepareWorkflow
      summary: Select a bounded workflow tool set without executing handlers
      tags:
      - runtime
      parameters:
      - $ref: '#/components/parameters/ContractVersion'
      - $ref: '#/components/parameters/Traceparent'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowView'
        '400':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: See shared/CONTRACT.md error registry; not all errors apply to
            every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected internal error; not automatically retryable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-required-key-scope: runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRequest'
...
  /v1/capabilities:
    get:
      operationId: getCapabilities
      summary: Negotiate Toolgate contract, not MCP protocol
      tags:
      - runtime
      parameters:
      - $ref: '#/components/parameters/ContractVersion'
      - $ref: '#/components/parameters/Traceparent'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capabilities'
        '400':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: See shared/CONTRACT.md error registry; not all errors apply to every route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected internal error; not automatically retryable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-required-key-scope: runtime
...
components:
  securitySchemes:
    ServiceKey:
      type: http
      scheme: bearer
      description: Opaque service credential bound to tenant/project/integration and scopes. NEVER forward
        the MCP user access token.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
    ContractVersion:
      name: X-Toolgate-Contract
      in: header
      required: true
      schema:
        const: '0.1'
    Traceparent:
      name: traceparent
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
  schemas:
    Actor:
      type: object
      properties:
        principal_ref: *id001
        workspace_ref: *id001
        authorization_revision:
          type: string
          minLength: 1
          maxLength: 128
        allowed_tool_ids:
          type: array
          items: *id001
          minItems: 0
          maxItems: 10000
          uniqueItems: true
      required:
      - principal_ref
      - workspace_ref
      - authorization_revision
      - allowed_tool_ids
      additionalProperties: false
      description: Built exclusively by trusted provider code. Not accepted from MCP arguments. Service
        tenant/project/integration derive from the API credential.
...
    ReadyOperation:
      type: object
      properties:
        operation_id: *id008
        revision: *id007
        catalog_id: *id001
        catalog_version: *id002
        expires_at: *id005
        status:
          const: ready
        selected_tool:
          $ref: '#/components/schemas/SelectedTool'
        arguments_digest: *id002
        next_tool:
          type: string
          enum:
          - execute_read_action
          - execute_write_action
      required:
      - operation_id
      - revision
      - catalog_id
      - catalog_version
      - expires_at
      - status
      - selected_tool
      - arguments_digest
      - next_tool
      additionalProperties: false
      description: Selection and schema validation succeeded. This is NOT permission to bypass provider
        authorization or approval.
...
    NeedsChoiceOperation:
      type: object
      properties:
        operation_id: *id008
        revision: *id007
        catalog_id: *id001
        catalog_version: *id002
        expires_at: *id005
        status:
          const: needs_choice
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
          minItems: 2
          maxItems: 5
        reason:
          type: string
          enum:
          - ambiguous_intent
          - low_confidence
        next_tool:
          const: prepare_action
      required:
      - operation_id
      - revision
      - catalog_id
      - catalog_version
      - expires_at
      - status
      - candidates
      - reason
      - next_tool
      additionalProperties: false
...
    OperationView:
      oneOf:
      - $ref: '#/components/schemas/ReadyOperation'
      - $ref: '#/components/schemas/NeedsArgumentsOperation'
      - $ref: '#/components/schemas/NeedsChoiceOperation'
      - $ref: '#/components/schemas/NoMatchOperation'
      - $ref: '#/components/schemas/CancelledOperation'
      - $ref: '#/components/schemas/IssuedOperation'
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
          pattern: ^[A-Z][A-Z0-9_]*$
        message_key:
          type: string
          minLength: 1
          maxLength: 160
        retryable:
          type: boolean
        request_id:
          type: string
          minLength: 1
          maxLength: 128
        issues:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
          minItems: 0
          maxItems: 128
      required:
      - code
      - message_key
      - retryable
      - request_id
      additionalProperties: false
    Error:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
      additionalProperties: false
...
    Capabilities:
      type: object
      properties:
        contract_version:
          const: 0.1.0
        schema_profiles:
          type: array
          items:
            const: tg-jsonschema-1
          minItems: 1
          maxItems: 1
        max_catalog_tools:
          const: 10000
        max_catalog_bytes:
          const: 16777216
        max_request_bytes:
          const: 262144
        operation_ttl_seconds:
          const: 900
        decision_ttl_seconds:
          const: 60
        mcp_execution_owner:
          const: provider-sdk
      required:
      - contract_version
      - schema_profiles
      - max_catalog_tools
      - max_catalog_bytes
      - max_request_bytes
      - operation_ttl_seconds
      - decision_ttl_seconds
      - mcp_execution_owner
      additionalProperties: false
    WorkflowRequest:
      type: object
      properties:
        catalog_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
        catalog_version:
          type: string
          pattern: ^[a-f0-9]{64}$
        actor:
          $ref: '#/components/schemas/Actor'
        intent:
          type: string
          minLength: 1
          maxLength: 8192
      required:
      - catalog_id
      - catalog_version
      - actor
      - intent
      additionalProperties: false
    WorkflowView:
      type: object
      additionalProperties: false
      required:
      - status
      - operations
      - candidate_count
      properties:
        status:
          type: string
          enum:
          - prepared
          - no_match
          - needs_refinement
        candidate_count:
          type: integer
          minimum: 0
          maximum: 64
        operations:
          type: array
          maxItems: 8
          items:
            $ref: '#/components/schemas/OperationView'